Make the math output prettier in the web UI
It uses MathML, but Electron/Chromium doesn't support it, so I used a fallback
CSS instead. :( Right now it doesn't look as good as native MathML
rendering. I'm not sure if I'll edit the CSS or try KaTeX (but I would have to
generate TeX automatically instead of MathML). MathJax seems like it's too
slow (need synchronous rendering ideally so it doesn't flicker when an equation
is added to the window).